% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_surv_func.R
\name{est_surv_func}
\alias{est_surv_func}
\title{Estimate survival functions for subgroups}
\usage{
est_surv_func(object, z_value, x_value = NULL, event_time, n.random = 1000)
}
\arguments{
\item{object}{an object of jmfhc_point_est.}

\item{z_value}{value(s) of the long-term baseline covariate(s) in the order specified at the argument beta_variable of jmfhc_point_est() function.}

\item{x_value}{value(s) of the short-term baseline covariate(s) in the order specified at the argument gamma_variable of jmfhc_point_est() function.
If gamma is set as NULL in the jmfhc_point_est() function, then there is no need to put values here. The default is NULL.}

\item{event_time}{the variable name of observed survival time in the data.}

\item{n.random}{the number of randomly generated random effects based on the estimated covariance matrix for random effects. The default is 1000.}
}
\value{
a data.frame with sorted survival times and the corresponding estimated survival probabilities.
}
\description{
Estimates the survival function for a subgroup of patients with specified characteristics.
}
\examples{
result_coef <- jmfhc_point_est(data=jmfhc_dat, event_time="event.time", event_status="event",
                               id="patient.id", beta_variable="trt", gamma_variable="trt",
                               fu_measure="measure", fu_time_variable="mes.times")
# estimated survival function for patient subgroup receiving treatment A
survival_trt0 <- est_surv_func(object=result_coef,z_value=0,x_value=0,event_time="event.time")
# estimated survival function for patient subgroup receiving treatment B
survival_trt1 <- est_surv_func(object=result_coef,z_value=1,x_value=1,event_time="event.time")
}
