% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcure.R
\name{estcure}
\alias{estcure}
\title{Estimate cure rates for subgroups}
\usage{
estcure(object, z_value)
}
\arguments{
\item{object}{an object of jmfhc_point_est() function.}

\item{z_value}{value(s) of the long-term baseline covariate(s) in the order specified at the argument beta_variable of jmfhc_point_est() function.}
}
\value{
an estimated cure rate
}
\description{
Estimates a cure rate for a subgroup of patients with specified characteristics.
}
\examples{
result_coef <- jmfhc_point_est(data=jmfhc_dat, event_time="event.time", event_status="event",
                               id="patient.id", beta_variable="trt", gamma_variable="trt",
                               fu_measure="measure", fu_time_variable="mes.times")
# cure rate for patient subgroup receiving treatment A (trt=0)
estcure(object=result_coef,z_value=0)
# cure rate for patient subgroup receiving treatment B (trt=1)
estcure(object=result_coef,z_value=1)
}
